refactor #33 : 탈퇴한 사용자일 경우 예외 처리 추가

+++ module_user/src/main/java/com/hanghae/module_user/user/service/UserService.java
@@ -98,6 +98,10 @@ public class UserService {
         User targetUser = userRepository.findById(userNum)
                 .orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
 
+        if (targetUser.getDeletedAt() != null) {
+            throw new CustomException(ErrorCode.DELETED_USER);
+        }
+
         targetUser.delete();
 
         return new UserResponse(targetUser.getUserNum(), targetUser.getName(), targetUser.getEmail());
